requirements.txt
numpy==2.4.4
pandas==3.0.2
pypdf==6.10.2
python-dateutil==2.9.0.post0
six==1.17.0
google-generativeai>=0.8.0
streamlit>=1.30.0
python-dotenv>=1.0.0
